app/requirements.txt
requests
fastapi
psycopg2-binary
uvicorn